refactor(harness): extraer workflow-dag y run-store de run-engine — fase 1 (04/T05)
Extracción sin reescritura, run-engine.cjs queda como fachada con la API
pública intacta (2.317 → 1.933 líneas). workflow-dag.cjs: funciones puras
del DAG (topologicalLevels, mergePayloads, getInputPayloads) con 7 tests
unitarios. run-store.cjs: persistencia de runs/steps/links y eventos
runs:updated/runs:step; el estado terminal de automations vuelve al
engine vía hook onTerminalAutomationStatus para evitar import circular.
Diagrama de módulos actualizado en docs/architecture/agent-runtime.md.
Fase 2 pendiente: workflow-executor y run-lifecycle.
